html-report: move code into html_helpers
Move functions from `html_report.py` into `html_helpers.py` to have
better separation between the code that assembles the overall report and
that which creates the HTML text itself.
